20130903 versie Dynamo-2 bijna definitief maar nog niet gecheckt of het wel werkt want op campus open

+++ Dynamo-HIA_UI/src/nl/rivm/emi/dynamo/data/types/atomic/base/AbstractClassIndex.java
@@ -20,7 +20,7 @@ abstract public class AbstractClassIndex extends FlexibleUpperLimitNumberRangeTy
 	final public Pattern matchPattern = Pattern.compile("^\\d*$");
 
 	static final protected Integer hardLowerLimit = new Integer(1);
-	static final protected Integer hardUpperLimit = new Integer(10);
+	static final protected Integer hardUpperLimit = new Integer(50);
 
 
 	/**
